requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.1
django-redis>=5.4
django-cors-headers>=4.3
django-filter>=24.0
python-dotenv>=1.0
bleach>=6.0
gunicorn>=22.0
# Phase 2: WebSocket + Channels
channels>=4.1
daphne>=4.1
channels-redis>=4.2
# Phase 3: Async Processing + Workers
celery[redis]>=5.3
django-anymail[sendgrid,amazon-ses]>=11.0
fcm-django>=2.0
pybreaker>=1.0
# Phase 6: Monitoring & Observability
django-prometheus>=2.3
python-json-logger>=2.0
django-guid>=3.4
opentelemetry-api>=1.20
opentelemetry-sdk>=1.20
opentelemetry-instrumentation-django>=0.44b0
opentelemetry-instrumentation-celery>=0.44b0
opentelemetry-exporter-jaeger>=1.20
